+++ lib/Actions/Package.php
@@ -6,7 +6,7 @@
  * by hand on every server.
  *
  * @package cwp7
- * @version 2.1.1
+ * @version 2.2.0
  * @author  Booysen Logistics <bradley@booysenlogistics.co.za>
  * @license MIT
  * @link    https://github.com/bradleygb/whmcs-cwp-module